+++ Reflection/src/utopia/reflection/component/swing/input/TextField.scala
@@ -115,18 +115,18 @@ object TextField
 	  * @param targetWidth The stack width for this field
 	  * @param document Document used for this field (default = plain document)
 	  * @param initialText Initially displayed text (default = no text)
-	  * @param prompt Prompt text displayed (optional)
+	  * @param prompt Prompt text displayed (default = empty = no prompt)
 	  * @param resultFilter A regex used for transforming field content (default = None)
 	  * @param context Component creation context
 	  * @return A new text field
 	  */
 	def contextual(targetWidth: StackLength, document: Document = new PlainDocument(), initialText: String = "",
-				   prompt: Option[LocalizedString] = None, resultFilter: Option[Regex] = None,
+				   prompt: LocalizedString = LocalizedString.empty, resultFilter: Option[Regex] = None,
 				   valuePointer: PointerWithEvents[Option[String]] = new PointerWithEvents(None))
 				  (implicit context: ButtonContextLike) =
 	{
 		val field = new TextField(targetWidth, context.textInsets.total / 2, context.font, document,
-			initialText, prompt.map { Prompt(_, context.promptFont, context.hintTextColor) }, context.textColor,
+			initialText, prompt.notEmpty.map { Prompt(_, context.promptFont, context.hintTextColor) }, context.textColor,
 			resultFilter, context.textAlignment, valuePointer)
 		field.background = context.buttonColor
 		field.setSelectionHighlight(if (context.colorScheme.secondary.contains(context.buttonColor))
@@ -139,11 +139,12 @@ object TextField
 	  * Creates a field that is used for writing positive integers. Uses component creation context.
 	  * @param targetWidth The stack width for this field
 	  * @param initialValue Initially displayed value (Default = None)
-	  * @param prompt Prompt text displayed, if any (Default = None)
+	  * @param prompt Prompt text displayed, if any (default = empty = no prompt)
 	  * @param context Component creation context (implicit)
 	  * @return A new text field
 	  */
-	def contextualForPositiveInts(targetWidth: StackLength, initialValue: Option[Int] = None, prompt: Option[LocalizedString] = None,
+	def contextualForPositiveInts(targetWidth: StackLength, initialValue: Option[Int] = None,
+	                              prompt: LocalizedString = LocalizedString.empty,
 								  valuePointer: PointerWithEvents[Option[String]] = new PointerWithEvents(None))
 								 (implicit context: ButtonContextLike) = contextual(targetWidth,
 		FilterDocument(Regex.digit, 10), initialValue.map { _.toString } getOrElse "", prompt, Some(Regex.numericPositive),
@@ -153,11 +154,12 @@ object TextField
 	  * Creates a field that is used for writing positive or negative integers. Uses component creation context.
 	  * @param targetWidth The stack width for this field
 	  * @param initialValue Initially displayed value (Default = None)
-	  * @param prompt Prompt text displayed, if any (Default = None)
+	  * @param prompt Prompt text displayed, if any (default = empty = no prompt)
 	  * @param context Component creation context (implicit)
 	  * @return A new text field
 	  */
-	def contextualForInts(targetWidth: StackLength, initialValue: Option[Int] = None, prompt: Option[LocalizedString] = None,
+	def contextualForInts(targetWidth: StackLength, initialValue: Option[Int] = None,
+	                      prompt: LocalizedString = LocalizedString.empty,
 						  valuePointer: PointerWithEvents[Option[String]] = new PointerWithEvents(None))
 						 (implicit context: ButtonContextLike) = contextual(targetWidth,
 		FilterDocument(Regex.numericParts, 11), initialValue.map { _.toString } getOrElse "", prompt, Some(Regex.numeric),
@@ -167,12 +169,12 @@ object TextField
 	  * Creates a field that is used for writing positive doubles. Uses component creation context.
 	  * @param targetWidth The stack width for this field
 	  * @param initialValue Initially displayed value (Default = None)
-	  * @param prompt Prompt text displayed, if any (Default = None)
+	  * @param prompt Prompt text displayed, if any (default = empty = no prompt)
 	  * @param context Component creation context (implicit)
 	  * @return A new text field
 	  */
 	def contextualForPositiveDoubles(targetWidth: StackLength, initialValue: Option[Double] = None,
-									 prompt: Option[LocalizedString] = None,
+									 prompt: LocalizedString = LocalizedString.empty,
 									 valuePointer: PointerWithEvents[Option[String]] = new PointerWithEvents(None))
 									(implicit context: ButtonContextLike) = contextual(targetWidth,
 		FilterDocument(Regex.decimalPositiveParts, 24), initialValue.map { _.toString } getOrElse "", prompt,
@@ -186,7 +188,8 @@ object TextField
 	  * @param context Component creation context (implicit)
 	  * @return A new text field
 	  */
-	def contextualForDoubles(targetWidth: StackLength, initialValue: Option[Double] = None, prompt: Option[LocalizedString] = None,
+	def contextualForDoubles(targetWidth: StackLength, initialValue: Option[Double] = None,
+	                         prompt: LocalizedString = LocalizedString.empty,
 							 valuePointer: PointerWithEvents[Option[String]] = new PointerWithEvents(None))
 							(implicit context: ButtonContextLike) = contextual(targetWidth,
 		FilterDocument(Regex.decimalParts, 24), initialValue.map { _.toString } getOrElse "", prompt,