+++ Reflection/test/utopia/reflection/test/MultiLineTextViewTest.scala
@@ -33,7 +33,7 @@ object MultiLineTextViewTest extends App
 		// Creates controls to bottom
 		val bottomRow = bc.forGrayFields.use { implicit fieldContext =>
 			// Creates text input
-			val textInput = TextField.contextual(standardWidth, prompt = Some("Type your own text and press enter"))
+			val textInput = TextField.contextual(standardWidth, prompt = "Type your own text and press enter")
 			textInput.addEnterListener { _.foreach { s => textView.text = (s: LocalString).localizationSkipped } }
 			
 			val alignSelect = JDropDownWrapper.contextual("Select Alignment", initialChoices = Alignment.values)